+++ src/parsing/scope.ml
@@ -648,7 +648,8 @@ let scope_rule :
       "Symbol %s has been declared constant, it cannot be used as the \
        head of a rewrite rule LHS." pr_sym.sym_name;
   if Timed.(!(pr_sym.sym_opaq) || (!(pr_sym.sym_def) <> None)) then
-    fatal p_lhs.pos "No rewriting rule can be added on an opaque symbol or a symbol already defined with ≔";
+    fatal p_lhs.pos "No rewriting rule can be added on an opaque symbol \
+                     or a symbol already defined with ≔";
   if pr_sym.sym_expo = Protec
     && ss.signature.sign_path <> pr_sym.sym_path then
     fatal p_lhs.pos "Cannot define rules on foreign protected symbols.";